minor changes to non wrath modules. Major changes to wrath specific cloned modules
The addon loads and tracks dailies/lockouts/currencies for wrath. Have not teseted alot of the functionallity i assume there is still bugs in areas of code having to do with features that arent in the wotlk game. (callings warfronts etc)

trying to keep the logic of the code similar to the source but just updating variable names and adding types for any future compatability. (and for my DX)