feat: Allow hiding of untracked currencies from the "Show All" Feature
by default, even when the currency is not selected in the options, the addon will generate a cell for all currencies if any alt has some of respective currency.

this feature adds an option to opt out of the default in aim to not clutter the tooltip on "Show All" states.